Isolate builders into files and implement attribute support
